+++ packages/typescript-estree/src/create-program/createWatchProgram.ts
@@ -174,7 +174,7 @@ function getProgramsForProjects(
       log('Found existing program for file. %s', filePath);
 
       updatedProgram =
-        updatedProgram || existingWatch.getProgram().getProgram();
+        updatedProgram ?? existingWatch.getProgram().getProgram();
       // sets parent pointers in source files
       updatedProgram.getTypeChecker();